pyproject.toml
[project]
name = "automatic-downlink"
version = "0.1.0"
description = "VLM-powered on-board satellite image triage for automatic downlink prioritization"
requires-python = ">=3.11"
dependencies = [
    "transformers>=4.57.0",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]
dashboard = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "jinja2>=3.1.0",
]
training = [
    "modal>=0.72.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
